# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API de Transferências e Usuários",
    "version": "1.0.0",
    "description": "API para login, registro, consulta de usuários e transferências."
  },
  "paths": {
    "/users/register": {
      "post": {
        "summary": "Registrar novo usuário",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": { "type": "string" },
                  "password": { "type": "string" },
                  "favorecido": { "type": "boolean" }
                },
                "required": ["username", "password"]
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Usuário registrado" },
          "400": { "description": "Erro de validação ou usuário duplicado" }
        }
      }
    },
    "/users/login": {
      "post": {
        "summary": "Login de usuário",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": { "type": "string" },
                  "password": { "type": "string" }
                },
                "required": ["username", "password"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Login bem-sucedido" },
          "401": { "description": "Credenciais inválidas" }
        }
      }
    },
    "/users": {
      "get": {
        "summary": "Consultar usuários",
        "responses": {
          "200": { "description": "Lista de usuários" }
        }
      }
    },
    "/transfer": {
      "post": {
        "summary": "Realizar transferência",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "from": { "type": "string" },
                  "to": { "type": "string" },
                  "amount": { "type": "number" }
                },
                "required": ["from", "to", "amount"]
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Transferência realizada" },
          "400": { "description": "Erro de validação ou regra de negócio" },
          "401": { "description": "Token não informado ou inválido" }
        }
      },
      "get": {
        "summary": "Consultar transferências",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "Lista de transferências" },
          "401": { "description": "Token não informado ou inválido" }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}